refactoring snapshot
