+++ src/OrleansRuntime/Catalog/ActivationData.cs
@@ -831,7 +831,7 @@ namespace Orleans.Runtime
 
         public override string ToString()
         {
-#if PQ_DEBUG
+#if DEBUG
             return String.Format("[Activation: {0}{1}{2}{3} {4} State={5}]",
                  Silo,
                  Grain,
@@ -872,7 +872,7 @@ namespace Orleans.Runtime
         {                  
             get
             {
-#if PQ_DEBUG
+#if DEBUG
                 return String.Format("[Activation: {0}{1}{2}{3} {4}]",
                      Silo,
                      Grain,
