+++ src/OrleansRuntime/Scheduler/PoliciedScheduler/SchedulingStrategies/WindowIDSchedulingStrategy.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Orleans.Runtime.Scheduler.Utility;
+using Orleans.Runtime.Scheduler.SchedulerUtility;
 
 namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
 {
