+++ src/OrleansRuntime/Scheduler/SchedulerUtility/SchedulerConstants.cs
@@ -1,6 +1,6 @@
 ﻿
 
-namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
+namespace Orleans.Runtime.Scheduler.SchedulerUtility
 {
     internal static class SchedulerConstants
     {
@@ -10,5 +10,6 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
         internal const int DEFAULT_TASK_QUANTUM_NUM_TASKS = 0;
         internal const long DEFAULT_DATAFLOW_SLA = 5000000;
         internal const long DEFAULT_WINODW_SIZE = 100000000;
+        internal const int MEASUREMENT_PERIOD_WORKITEM_COUNT = 100;
     }
 }
